refactor: delete app nav host's scaffold layout
