+++ core/ui/src/main/res/drawable/ic_calendar.png
