+++ core/ui/src/main/res/drawable/ic_home.png
